Add data sending feature and update version
A new functionality has been added to send visit parameter data to the server; this function constructs a POST request and sends this to the specified endpoint. In addition to this, an error-safe 'delete' operation for removing specific data from local storage has been implemented. The version in package.json has also been updated to reflect these changes.